pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "math-mirror-ai"
version = "0.0.1"
description = "Agentic math framework. Trained only on math. Does only math."
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "sympy>=1.12",
]

[project.optional-dependencies]
jax = ["jax[cpu]", "flax"]
mcp = ["arxiv>=2.1", "openai>=1.0", "anthropic>=0.18", "google-generativeai>=0.4", "aiohttp>=3.9"]
dev = ["pytest"]
